feat(EctoAdmin.Config): module for fetching host application config
This class when initialized requires an `otp_app` atom and a module
name. In this case `otp_app` will be the name of the host application.
The `module` argument will always be `EctoAdmin` which is used as the
config key. The host application is expected to set this config in a
config file by using a config file such as:

    config :admin_test, EctoAdmin,
      repo:   AdminTest.Repo,
      models: [users: AdminTest.User]

The config module is a GenServer which is started when the application
starts.