feat(EctoAdmin.Router): serve all admin requests
The EctoAdmin.Rotuer module servers the requests. It is expected that
this will be mounted in the host application. It currently tries to find
the resource based on the config and then if found, passes it off to
EctoAdmin.Resource

The response is wrapped inside of an EEx template that is currently
styled using Twitter Bootstrap.