feat(EctoAdmin.Resource): module to render the HTML for resources
The Resource module exposes a single render function which can be used
to render an action for a particular resource. Currently only the index
resource route matches. It is rendered into an EEx template.

There is a special case where an Ecto.DateTime struct cannot be rendered
as a string. For this reason, each instance is mapped to replace the
Ecto.DateTime struct with a string. To facilitate this, the instance
struct is converted to a Keyword list so that each field can be
enumerated.