refactor(EctoAdmin.Resource): Ecto.DateTime logic now in EctoAdmin.View
A new module EctoAdmin.View has been created as a view helper for the
templates. The logic that was previously used to stringify the
Ecto.DateTime fields has been moved into this module. This means that it
is no longer required to loop over the fields in the struct to check
their values, this is now checked at the template level with a call to
`EctoAdmin.View.field`